+++ lib/networks/network_context.go
@@ -15,6 +15,7 @@ import (
 	"github.com/sirupsen/logrus"
 	"os"
 	"path/filepath"
+	"sync"
 )
 
 const (
@@ -26,6 +27,9 @@ const (
 )
 
 type NetworkContext struct {
+	// Mutex, to make this thread-safe
+	mutex *sync.Mutex
+
 	kurtosisService kurtosis_service.KurtosisService
 
 	// The dirpath ON THE SUITE CONTAINER where the suite execution volume is mounted
@@ -54,6 +58,7 @@ func NewNetworkContext(
 		suiteExecutionVolumeDirpath string,
 		servicesRelativeDirpath string) *NetworkContext {
 	return &NetworkContext{
+		mutex: &sync.Mutex{},
 		kurtosisService: kurtosisService,
 		suiteExecutionVolumeDirpath: suiteExecutionVolumeDirpath,
 		servicesRelativeDirpath: servicesRelativeDirpath,
@@ -63,6 +68,9 @@ func NewNetworkContext(
 
 // Gets the number of nodes in the network
 func (networkCtx *NetworkContext) GetSize() int {
+	networkCtx.mutex.Lock()
+	defer networkCtx.mutex.Unlock()
+
 	return len(networkCtx.services)
 }
 
@@ -82,6 +90,7 @@ Return:
 func (networkCtx *NetworkContext) AddService(
 		serviceId services.ServiceID,
 		initializer services.DockerContainerInitializer) (services.Service, services.AvailabilityChecker, error) {
+	// Mutex locked directly inside (we can't lock the mutex here because Go mutexes aren't reentrant)
 	service, availabilityChecker, err := networkCtx.AddServiceToPartition(
 		serviceId,
 		defaultPartitionId,
@@ -110,6 +119,9 @@ func (networkCtx *NetworkContext) AddServiceToPartition(
 		serviceId services.ServiceID,
 		partitionId PartitionID,
 		initializer services.DockerContainerInitializer) (services.Service, services.AvailabilityChecker, error) {
+	networkCtx.mutex.Lock()
+	defer networkCtx.mutex.Unlock()
+
 	if _, exists := networkCtx.services[serviceId]; exists {
 		return nil, nil, stacktrace.NewError("Service ID %s already exists in the network", serviceId)
 	}
@@ -193,6 +205,9 @@ func (networkCtx *NetworkContext) AddServiceToPartition(
 Gets the node information for the service with the given service ID.
 */
 func (networkCtx *NetworkContext) GetService(serviceId services.ServiceID) (services.Service, error) {
+	networkCtx.mutex.Lock()
+	defer networkCtx.mutex.Unlock()
+
 	service, found := networkCtx.services[serviceId]
 	if !found {
 		return nil, stacktrace.NewError("No service with ID '%v' exists in the network", serviceId)
@@ -205,6 +220,9 @@ func (networkCtx *NetworkContext) GetService(serviceId services.ServiceID) (serv
 Stops the container with the given service ID, and removes it from the network.
 */
 func (networkCtx *NetworkContext) RemoveService(serviceId services.ServiceID, containerStopTimeoutSeconds int) error {
+	networkCtx.mutex.Lock()
+	defer networkCtx.mutex.Unlock()
+
 	_, found := networkCtx.services[serviceId]
 	if !found {
 		return stacktrace.NewError("No service with ID %v found", serviceId)
@@ -232,6 +250,7 @@ Args:
 		during a repartition then traffic between them will be blocked by default
  */
 func (networkCtx NetworkContext) GetRepartitionerBuilder(isDefaultPartitionConnectionBlocked bool) *RepartitionerBuilder {
+	// This function doesn't need a mutex lock because (as of 2020-12-28) it doesn't touch internal state whatsoever
 	return newRepartitionerBuilder(isDefaultPartitionConnectionBlocked)
 }
 
@@ -240,6 +259,9 @@ Repartitions the network using the given repartitioner. A repartitioner builder 
 	NewRepartitionerBuilder method of this network context object.
  */
 func (networkCtx *NetworkContext) RepartitionNetwork(repartitioner *Repartitioner) error {
+	networkCtx.mutex.Lock()
+	defer networkCtx.mutex.Unlock()
+
 	partitionServices := map[string]map[string]bool{}
 	for partitionId, serviceIdSet := range repartitioner.partitionServices {
 		serviceIdStrPseudoSet := map[string]bool{}