+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 * Add the `Service.GetServiceID` method
 * Renamed `DockerContainerInitializer.GetServiceFromIp` -> `GetService`, and passed in the `ServiceID` as a new first argument
     * All `Service` implementations should have their constructors modified to store this new argument
+* Implemented the ability to partition test networks! This brought several changes:
+    * Added a `GetTestConfiguration` function to the `Test` interface, which allows tests to declare certain types of functionality (like network partitioning)
+    * Added `NetworkPartitionTest` to test the new network partitioning functionality
+    * Made `NetworkContext` thread-safe
 
 ## 1.4.1
 * Point all old `kurtosis-docs` references to `docs.kurtosistech.com`